Remove legacy C constructions
The C99 standard removed several backward compatibility features. Switch
to their modern counterpart to prepare for future compilers.

+++ scilab/modules/javasci/src/c/javasci2_helper.h
@@ -277,7 +277,7 @@ int putLong(char* variableName, long *variable, int nbRow, int nbCol);
  * @param[out] nbCol Number of cols
  * @return The matrix of unsigned long (column sorted)
  */
-unsigned long * getUnsignedLong(char* variableName, int *nbRow, int *nbCol);
+unsigned long long* getUnsignedLong(char* variableName, int *nbRow, int *nbCol);
 
 /**
  * Set a matrix of unsigned long (int64)
@@ -288,7 +288,7 @@ unsigned long * getUnsignedLong(char* variableName, int *nbRow, int *nbCol);
  * @param nbCol Number of cols
  * @return 0 if successfull, != 0 otherwise
  */
-int putUnsignedLong(char* variableName, unsigned long *variable, int nbRow, int nbCol);
+int putUnsignedLong(char* variableName, unsigned long long* variable, int nbRow, int nbCol);
 
 /**
  * Return a matrix of string
